+++ backend/src/baserow/contrib/database/apps.py
@@ -247,6 +247,8 @@ class DatabaseConfig(AppConfig):
             SingleSelectNotEqualViewFilterType,
             LinkRowHasViewFilterType,
             LinkRowHasNotViewFilterType,
+            LinkRowContainsViewFilterType,
+            LinkRowNotContainsViewFilterType,
             MultipleSelectHasViewFilterType,
             MultipleSelectHasNotViewFilterType,
             LengthIsLowerThanViewFilterType,
@@ -274,6 +276,8 @@ class DatabaseConfig(AppConfig):
         view_filter_type_registry.register(SingleSelectNotEqualViewFilterType())
         view_filter_type_registry.register(LinkRowHasViewFilterType())
         view_filter_type_registry.register(LinkRowHasNotViewFilterType())
+        view_filter_type_registry.register(LinkRowContainsViewFilterType())
+        view_filter_type_registry.register(LinkRowNotContainsViewFilterType())
         view_filter_type_registry.register(BooleanViewFilterType())
         view_filter_type_registry.register(EmptyViewFilterType())
         view_filter_type_registry.register(NotEmptyViewFilterType())