+++ web-frontend/modules/database/plugin.js
@@ -51,6 +51,8 @@ import {
   LinkRowHasNotFilterType,
   MultipleSelectHasFilterType,
   MultipleSelectHasNotFilterType,
+  LinkRowContainsFilterType,
+  LinkRowNotContainsFilterType,
 } from '@baserow/modules/database/viewFilters'
 import {
   CSVImporterType,
@@ -267,6 +269,11 @@ export default (context) => {
   app.$registry.register('viewFilter', new BooleanViewFilterType(context))
   app.$registry.register('viewFilter', new LinkRowHasFilterType(context))
   app.$registry.register('viewFilter', new LinkRowHasNotFilterType(context))
+  app.$registry.register('viewFilter', new LinkRowContainsFilterType(context))
+  app.$registry.register(
+    'viewFilter',
+    new LinkRowNotContainsFilterType(context)
+  )
   app.$registry.register('viewFilter', new MultipleSelectHasFilterType(context))
   app.$registry.register(
     'viewFilter',